requirements.txt
langchain==0.3.27
langgraph==0.6.5
langchain-openai==0.3.30
langchain-community==0.3.27
python-dotenv==1.1.1
langchain-tavily==0.2.11
wikipedia
arxiv
streamlit